refactor code_collaborative_analysis.py
