+++ src/analysis/code_collaborative/code_collaborative_analysis_helper.py
@@ -9,6 +9,11 @@ import json
 
 from src.utils.extension_catalog import get_languages_for_extension
 from src.utils.helpers import ensure_table 
+from src.db.git_identities import (
+    ensure_user_github_table,
+    load_user_github,
+    save_user_github,
+)
 # zip_paths stays in the main file because only the entrypoint needs it
 
 import re
@@ -150,53 +155,6 @@ def _match_git_path(db_path: str, git_paths: set[str]) -> str | None:
             return gp
     return None
 
-
-
-# ------------------------------------------------------------
-# 2. user_github table + identity
-# ------------------------------------------------------------
-def ensure_user_github_table(conn: sqlite3.Connection) -> None:
-    ensure_table(
-        conn,
-        "user_github",
-        """
-        CREATE TABLE IF NOT EXISTS user_github (
-            user_id     INTEGER NOT NULL,
-            email       TEXT,
-            name        TEXT,
-            created_at  TEXT DEFAULT (datetime('now')),
-            UNIQUE(user_id, email, name)
-        )
-        """,
-    )
-
-
-def load_user_github(conn: sqlite3.Connection, user_id: int) -> Dict[str, set]:
-    emails, names = set(), set()
-    cur = conn.execute("SELECT email, name FROM user_github WHERE user_id = ?", (user_id,))
-    for em, nm in cur.fetchall():
-        if em:
-            emails.add(em.strip().lower())
-        if nm:
-            names.add(nm.strip().lower())
-    return {"emails": emails, "names": names}
-
-
-def save_user_github(conn: sqlite3.Connection, user_id: int, emails: List[str], names: List[str]) -> None:
-    cur = conn.cursor()
-    for em in set(e.strip().lower() for e in emails if e.strip()):
-        cur.execute(
-            "INSERT OR IGNORE INTO user_github(user_id, email, name) VALUES (?, ?, NULL)",
-            (user_id, em),
-        )
-    for nm in set(n.strip() for n in names if n.strip()):
-        cur.execute(
-            "INSERT OR IGNORE INTO user_github(user_id, email, name) VALUES (?, NULL, ?)",
-            (user_id, nm),
-        )
-    conn.commit()
-    
-    print("\nSaved your identity for future runs")
 
 
 def collect_repo_authors(repo_dir: str) -> List[Tuple[str, str, int]]:
